Make forced duration floating point as well

+++ src/client/java/mchorse/bbs_mod/ui/framework/elements/input/keyframes/UIKeyframeSheet.java
@@ -79,7 +79,7 @@ public class UIKeyframeSheet
         }
     }
 
-    public void setDuration(int duration)
+    public void setDuration(float duration)
     {
         for (Keyframe keyframe : this.selection.getSelected())
         {
